+++ src/App.js
@@ -6,6 +6,7 @@ import GlobalStyle from "./styles/GlobalStyle";
 import Routing from "./Routing";
 import {UserContext} from "./context/UserContext";
 import {ThemeContext} from "./context/ThemeContext";
+import { init } from "./utils/Web3Client";
 
 import Web3 from 'web3'
 import {user1} from "./utils/FakeBackend";
@@ -19,41 +20,15 @@ const App = () => {
     const [loading, setLoading] = useState(true);
     const [constructorHasRun, setConstructorHasRun] = useState(false);
 
-    const loadWeb3 = async () => {
-        if (window.ethereum) {
-            window.web3 = new Web3(window.ethereum)
-            await window.ethereum.enable()
-        }
-        else if (window.web3) {
-            window.web3 = new Web3(window.web3.currentProvider)
-        }
-        else {
-            window.alert('Non-Ethereum browser detected. You should consider trying MetaMask!')
-        }
-    }
-
-    const loadAddress = async () => {
-        const web3 = window.web3
-        // make sure app is connected to blockchain
-        const accounts = await web3.eth.getAccounts()
-        if (typeof accounts[0] === 'string' && accounts[0] !== '') {
-            toast.success(`app login success ${accounts[0]}`);
-            setUser({
-                "address" : accounts[0]
-            })
-        }
-    }
-
-    const fetchData = async () =>{
+    useEffect(async () => {
         if (constructorHasRun) return;
-        await loadWeb3();
-        await loadAddress();
-        setLoading(false)
-        setConstructorHasRun(true);
-    }
 
-    useEffect(() => {
-        fetchData();
+        setUser({
+            "address" : await init()
+        });
+        
+        setLoading(false);
+        setConstructorHasRun(true);
     }, [])
     
     if (loading) {
